mv conditionalattention to utils
